requirements.txt
PyYAML>=5.0
paramiko>=2.11.0
requests>=2.20
flask>=2.0
flask-socketio>=5.0
pydantic>=2.0

# Optional backends
# proxmoxer>=2.0    # Enable for REST API backend (backend: api)
# asyncssh>=2.14    # Enable for async SSH operations
